# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TACHIKOMA-OS Backend API",
    "description": "Central orchestrator API for the TACHIKOMA-OS AI ecosystem. Provides LLM gateway, GraphRAG memory, chat, agent tools, and data layer for microservices.",
    "contact": {
      "name": "TACHIKOMA-OS Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin/graph/export": {
      "get": {
        "tags": [
          "Graph"
        ],
        "summary": "=============================================================================",
        "description": "Export full graph (for visualization)\n=============================================================================\nGET /api/admin/graph/export\n=============================================================================",
        "operationId": "export_graph",
        "responses": {
          "200": {
            "description": "Full graph export",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphExportDto"
                }
              }
            }
          },
          "500": {
            "description": "Export error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/graph/stats": {
      "get": {
        "tags": [
          "Graph"
        ],
        "summary": "=============================================================================",
        "description": "Get graph statistics\n=============================================================================\nGET /api/admin/graph/stats\n=============================================================================",
        "operationId": "get_graph_stats",
        "responses": {
          "200": {
            "description": "Graph statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphStatsDto"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memories": {
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "GET /api/memories",
        "operationId": "list_memories",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-based)",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of memories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemoryDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "POST /api/memories",
        "operationId": "create_memory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Memory created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDto"
                }
              }
            }
          },
          "500": {
            "description": "Create error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/memories/search": {
      "post": {
        "tags": [
          "Memory"
        ],
        "summary": "POST /api/memories/search",
        "operationId": "search_memories",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResultDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Search error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/memories/{id}": {
      "get": {
        "tags": [
          "Memory"
        ],
        "summary": "GET /api/memories/:id",
        "operationId": "get_memory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memory found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Memory"
        ],
        "summary": "DELETE /api/memories/:id",
        "operationId": "delete_memory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Memory deleted"
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Delete error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Memory"
        ],
        "summary": "PATCH /api/memories/:id",
        "operationId": "update_memory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Memory updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Update error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateMemoryRequest": {
        "type": "object",
        "description": "Create memory request",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "description": "Memory content"
          },
          "importance_score": {
            "type": "number",
            "format": "double",
            "description": "Optional importance score",
            "nullable": true
          },
          "memory_type": {
            "type": "string",
            "description": "Memory type"
          },
          "metadata": {
            "description": "Optional metadata",
            "nullable": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "API error response",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "description": "Optional details",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "GraphExportDto": {
        "type": "object",
        "description": "Full graph export response",
        "required": [
          "nodes",
          "edges",
          "exported_at"
        ],
        "properties": {
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphEdgeDto"
            },
            "description": "All edges"
          },
          "exported_at": {
            "type": "string",
            "description": "Export timestamp"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryDto"
            },
            "description": "All nodes"
          }
        }
      },
      "GraphStatsDto": {
        "type": "object",
        "description": "Graph statistics response",
        "required": [
          "total_nodes",
          "total_edges",
          "nodes_by_type",
          "edges_by_type",
          "avg_connections"
        ],
        "properties": {
          "avg_connections": {
            "type": "number",
            "format": "double",
            "description": "Average connections per node"
          },
          "edges_by_type": {
            "type": "object",
            "description": "Edges by type",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "nodes_by_type": {
            "type": "object",
            "description": "Nodes by type",
            "additionalProperties": {
              "type": "integer",
              "minimum": 0
            }
          },
          "total_edges": {
            "type": "integer",
            "description": "Total number of edges",
            "minimum": 0
          },
          "total_nodes": {
            "type": "integer",
            "description": "Total number of nodes",
            "minimum": 0
          }
        }
      },
...
      "MemoryDto": {
        "type": "object",
        "description": "=============================================================================\nMemory DTOs\n=============================================================================\nMemory node response",
        "required": [
          "id",
          "content",
          "memory_type",
          "importance_score",
          "created_at",
          "updated_at",
          "access_count"
        ],
        "properties": {
          "access_count": {
            "type": "integer",
            "format": "int64",
            "description": "Access count",
            "minimum": 0
          },
          "content": {
            "type": "string",
            "description": "Memory content"
          },
          "created_at": {
            "type": "string",
            "description": "Created timestamp"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "importance_score": {
            "type": "number",
            "format": "double",
            "description": "Importance score (0.0 - 1.0)"
          },
          "memory_type": {
            "type": "string",
            "description": "Memory type"
          },
          "metadata": {
            "description": "Metadata"
          },
          "updated_at": {
            "type": "string",
            "description": "Updated timestamp"
          }
        }
      },
...
      "SearchResultDto": {
        "type": "object",
        "description": "Search result",
        "required": [
          "memory",
          "similarity"
        ],
        "properties": {
          "memory": {
            "$ref": "#/components/schemas/MemoryDto"
          },
          "similarity": {
            "type": "number",
            "format": "double",
            "description": "Similarity score"
          }
        }
      },
      "SemanticSearchRequest": {
        "type": "object",
        "description": "Semantic search request",
        "required": [
          "query"
        ],
        "properties": {
          "limit": {
            "type": "integer",
            "description": "Maximum results to return",
            "minimum": 0
          },
          "memory_types": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional memory type filter",
            "nullable": true
          },
          "min_similarity": {
            "type": "number",
            "format": "double",
            "description": "Minimum similarity score (0.0 - 1.0)"
          },
          "query": {
            "type": "string",
            "description": "Search query"
          }
        }
      },
...
    }
  },
...
}